Fix a bug in DotCmp (Similar to the one in SqEuclidean)

+++ libEM/cmp.cpp
@@ -300,7 +300,7 @@ float DotCmp::cmp(EMData* image, EMData* with) const
 			}
 			result += part;
 		}
-		n = nx*ny*nz*nx*ny*nz;
+		n = (long int)nx*(long int)ny*(long int)nz*(long int)nx*(long int)ny*(long int)nz;
 		
 		}else{ //This 3D code is incorrect, but it is the best I can do now 01/09/06 PAP
 		int ky, kz;
@@ -318,11 +318,11 @@ float DotCmp::cmp(EMData* image, EMData* with) const
 				}
 			}
 		}
-		n = nx*ny*nz*nx*ny*nz/2;
+		n = (long int)nx*(long int)ny*(long int)nz*(long int)nx*(long int)ny*(long int)nz/2;
 		}
 		result /= n;
 	} else {
-		long totsize = image->get_xsize() * image->get_ysize() * image->get_zsize();
+		long totsize = (long int)image->get_xsize() * (long int)image->get_ysize() * (long int)image->get_zsize();
 
 		double square_sum1 = 0., square_sum2 = 0.;
